+++ src/hooks/use-events.tsx
@@ -17,3 +17,12 @@ export const useKeyDown = (handleKeyDown: (event: KeyboardEvent) => void) => {
     };
   });
 };
+
+export const useScroll = (handleScrollEvent: (event: Event) => void) => {
+  useEffect(() => {
+    document.addEventListener("scroll", handleScrollEvent);
+    return () => {
+      document.removeEventListener("scroll", handleScrollEvent);
+    };
+  });
+};
